blender: Add Blender import mode
The Blender import mode can be used to view track in 3d viewport as
accurately as possible. For example, this mode arranges the shader nodes
to multiply albedo color with vertex shading.
This mode must never be used when the intent is to export the track to
GLTF, or other formats.